Move pyenv.ts
